requirements.txt
astroid==3.3.6
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.4.0
click==8.1.7
dill==0.3.9
exceptiongroup==1.2.2
greenlet>=3.1.1, <=4.0.0
hpy==0.9.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
lxml==5.3.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
progress==1.6
pycparser==2.22
pylint==3.3.2
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
readline==6.2.4.1
requests==2.32.3
ruff==0.8.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
